requirements.txt
Django==2.1.4
django-cors-headers==2.3.0
djangorestframework==3.9.0
Pillow==5.3.0
pytz==2018.7
psycopg2-binary==2.7.6.1